requirements.txt
Django==2.2.4
djangorestframework==3.10.2
gunicorn==19.9.0
mohawk==1.0.0
psycopg2-binary==2.8.3
requests==2.22.0
